+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 ## [Unreleased]
 ### Changed
  - All endpoints with a known operation verb (like `read` or `update`) now have an auto-generated operation id (`openapi` feature only)
+ - Endpoint macros now place the rustdoc into the operation's description (`openapi` feature only)
 
 ## [0.3.0] - 2021-03-21
 ### Changed