+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](https://semver.org).
 
 ## [Unreleased]
 ### Changed
+ - Rename endpoint macro `change` to `update` and `remove` to `delete`
  - All endpoints with a known operation verb (like `read` or `update`) now have an auto-generated operation id (`openapi` feature only)
  - Endpoint macros now place the rustdoc into the operation's description (`openapi` feature only)
  - Update `openapi_type` crate to 0.2.0